fix checking ack/from

+++ src/main/java/ru/mail/polis/kislichenko/RequestDelete.java
@@ -38,25 +38,24 @@ public class RequestDelete {
 
         for (int i = 0; goodReplics < from && i < hosts.length; i++) {
             if (ports[i] == myPort) {
-                if (dao.checkId(id)) {
-                    dao.delete(id);
-                    goodReplics++;
-                }
+                if (dao.checkId(id)) dao.delete(id);
+                goodReplics++;
                 continue;
             }
             int code;
 
             HttpResponse tmpStatus;
             try {
-                tmpStatus = Request.Delete(URLCreating.urlEntity(ports[i], id)).execute().returnResponse();
+                tmpStatus = Request.Delete(URLCreating.urlNodesIdInterior(ports[i], id)).execute().returnResponse();
             } catch (IOException e) {
                 continue;
             }
             code = tmpStatus.getStatusLine().getStatusCode();
 
-            if (code == 202) goodReplics++;
+            if (code ==202||code==404) goodReplics++;
 
         }
+        System.out.println("goodReplics: " + goodReplics);
         if (goodReplics >= ack) http.sendResponseHeaders(202, 0);
         else http.sendResponseHeaders(504, 0);
